just finished adding a translation function trans(msg) to echarts() and translated the text of many of the tooltip popups and other labels/text displayed on the page from Chinese to English.
I dont really know Chinese so the translations might not be perfect but its a good start.

The program now checks the browser language when it loads and if the language
happens to be 'en-US' it will return a translated phrase, if its not 'en-US' 
it returns the original text (Chinese).

I did it like this (keeping all the Chinese text in original spots and translating 
to english only if we're sure we got a en-US browser) because the developers clearly
use Chinese and it would probably be inconvenientand ONLY if en-US will it change) 
because as the program is obviously developed by people who do use Chinese it would
probably make things difficult if the default text wasnt in thier language.

Only 20 Phrases have been Translated so far but the function should provide an easy
way to quickly expand adding additional translations (and languages) as needed.