fix #2932, #2723 auto switch for tooltip trigger and multi tooltips in one echart instance
1. Add auto switch for tooltip. tooltip.trigger is discarded.
It can be switched in _tryShow().
If e.target.dataIndex exists, the type of trigger source is item or
showTip/showTipEx event.
Then check triggerAxis if the index of serie is not in the array, the
trigger source is item.

2. Add multi tooltips support.
Add showTipEx/hideTipEx and other code to avoid recursively dispatching
events.
Add viewId to get tootip index in option.
Users can show other tooltips (except tooltip[0]) by dispatch showTipEx
action in showTip event of tooltip[0].

Example is in #2932