feat(line)!: support both series and item delay with animationDelay callbacks
BREAKING CHANGE: May not be compatible with 4.x.
Since default symbol animation with line charts has been changed in 5.0 to be
displayed when the clipPath of the line goes through the data point,
animationDelay callback now has two meanings:
animationDelay: function (i) { return ... }
(1) If `i` is `null`, the callback should return the delay of the overall
    series animation, that is, the delay before the line clipPath starts
    changing.
(2) If `i` is a `number`, the callback should return the delay of the certain
    delay of each data item.
In ECharts 4.x, however, since the label animation shows along with the line
animation, only the `number` form is used.