feat(pie): Add new roseType value 'scaledArea' and corresponding test cases #15674
This commit adds a new value 'scaledArea' to the roseType option in rose chart. When set to 'scaledArea', the radius is calculated based on the square root, and the ratio of the actual value is proportional to the ratio of the area occupied by different items. Additionally, new test cases have been added to demonstrate the effects of different roseType values.

My testing shows that when the value is negative, the pie chart will lack that specific item, but it is unclear how the code handles it.